Refactor the processor module.
Have the processor class in a separate module. Also removed the textrank calculation as that is not need (for now) in the following mining step. The filter function is removed and will be replaced by the built-in Orange widget ow selector.
